# docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  handler.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  model.Article:
    properties:
      categories:
        items:
          $ref: '#/definitions/model.Category'
        type: array
      comment_count:
        type: integer
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      content:
        type: string
      created_at:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      read_count:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      updatedAt:
        type: string
    type: object
  model.Category:
    properties:
      articles:
        items:
          $ref: '#/definitions/model.Article'
        type: array
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  model.Comment:
    properties:
      article:
        $ref: '#/definitions/model.Article'
      article_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      updated_at:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
  model.User:
    properties:
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      created_at:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      password:
        maxLength: 32
        minLength: 4
        type: string
      updatedAt:
        type: string
      username:
        maxLength: 12
        minLength: 4
        type: string
    required:
    - email
    - password
    - username
    type: object
  utils.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /api/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login Information
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Login a user
      tags:
      - auth
  /api/user:
    post:
      consumes:
      - application/json
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create a user
      tags:
      - user
  /api/user/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Delete a user
      tags:
      - user
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get a user
      tags:
      - user
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update a user
      tags:
      - user
  /api/user/{id}/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: New Password
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update a user's password
      tags:
      - user
  /api/users:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page Size
        in: query
        name: page_size
        type: integer
      - description: Page Number
        in: query
        name: page_num
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List users
      tags:
      - user
  /api/users/{username}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Page Size
        in: query
        name: page_size
        type: integer
      - description: Page Number
        in: query
        name: page_num
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List users by username
      tags:
      - user
swagger: "2.0"